LayoutTests/fast/events/spatial-navigation/snav-iframe-no-scrollable-content-expected.txt

a


a
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "1"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "2"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "3"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "7"
PASS document.defaultView.frames[1].document.activeElement.getAttribute("id") is "6"
PASS document.activeElement.getAttribute("id") is "end"
PASS document.defaultView.frames[1].document.activeElement.getAttribute("id") is "4"
PASS document.defaultView.frames[1].document.activeElement.getAttribute("id") is "5"
PASS document.defaultView.frames[1].document.activeElement.getAttribute("id") is "6"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "7"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "3"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "2"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "1"
